Moved acl and meta Link headers to the generic handler
