+++ lib/handlers/get.js
@@ -57,20 +57,6 @@ function get(req, res, includeBody) {
         debug('HEAD -- ' + req.originalUrl);
     }
 
-    // Add ACL and Meta Link in header
-    var aclLink = utils.getResourceLink(
-        filename, uri,
-        ldp.root, ldp.suffixAcl,
-        ldp.suffixMeta);
-
-    var metaLink = utils.getResourceLink(
-        filename, uri,
-        ldp.root, ldp.suffixMeta,
-        ldp.suffixAcl);
-
-    header.addLink(res, aclLink, 'acl');
-    header.addLink(res, metaLink, 'describedBy');
-
     // Get resource or container
     ldp.get(filename, uri, includeBody, function(err, data, container) {
