+++ lib/header.js
@@ -4,6 +4,7 @@ var path = require('path');
 var S = require('string');
 var Negotiator = require('negotiator');
 var metadata = require('./metadata.js');
+var ldp = require('./ldp.js');
 var utils = require('./utils.js');
 var ldpVocab = require('./vocab/ldp.js');
 
@@ -30,6 +31,7 @@ function addLinks(res, fileMetadata) {
 }
 
 function linksHandler(req, res, next) {
+    var uri = utils.uriBase(req);
     var ldp = req.app.locals.ldp;
     var filename = utils.uriToFilename(req.url, ldp.root);
 
@@ -45,6 +47,22 @@ function linksHandler(req, res, next) {
     } else {
         fileMetadata.isResource = true;
     }
+
+    // Add ACL and Meta Link in header
+    var aclLink = utils.getResourceLink(
+        filename, uri,
+        ldp.root, ldp.suffixAcl,
+        ldp.suffixMeta);
+
+    var metaLink = utils.getResourceLink(
+        filename, uri,
+        ldp.root, ldp.suffixMeta,
+        ldp.suffixAcl);
+
+    addLink(res, aclLink, 'acl');
+    addLink(res, metaLink, 'describedBy');
+
+
     addLinks(res, fileMetadata);
     next();
 }
